Domain name changes: yegor256 -> artipie
