+++ pom.xml
@@ -29,8 +29,8 @@ SOFTWARE.
     <artifactId>parent</artifactId>
     <version>0.49.5</version>
   </parent>
-  <groupId>com.yegor256</groupId>
-  <artifactId>rpm-files</artifactId>
+  <groupId>com.artipie</groupId>
+  <artifactId>rpm-adapter</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
   <name>rpm-files</name>
