+++ src/main/java/com/artipie/rpm/Rpm.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.yegor256.rpm;
+package com.artipie.rpm;
 
 import com.yegor256.asto.Storage;
 import java.io.IOException;
