+++ src/test/java/com/artipie/rpm/RpmITCase.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.yegor256.rpm;
+package com.artipie.rpm;
 
 import com.jcabi.log.Logger;
 import com.yegor256.asto.Storage;
